Refactored custom fields, fixed lots of issues with loans, refactored disbursement creation form
