+++ assets/js/app/manage/ManagerContainer.jsx
@@ -29,33 +29,8 @@ const ManagerContainer = React.createClass({
     return (
       <Grid>
         <Row>
-          <Col sm={2}>
-            <Nav bsStyle="pills" stacked activeKey={this.state.activeKey} onSelect={this.handleSelect}>
-              <LinkContainer to="/app/manage/create-item">
-                <NavItem eventKey={1}>Create Items</NavItem>
-              </LinkContainer>
-              <LinkContainer to="/app/manage/tags">
-                <NavItem eventKey={7}>Tags</NavItem>
-              </LinkContainer>
-              { this.getCustomFieldsLink(this.props.route.admin.is_superuser) }
-              <LinkContainer to="/app/manage/disburse">
-                <NavItem eventKey={3}>Disbursement</NavItem>
-              </LinkContainer>
-              <LinkContainer to="/app/manage/requests">
-                <NavItem eventKey={4}>Requests</NavItem>
-              </LinkContainer>
-              <LinkContainer to="/app/manage/transactions">
-                <NavItem eventKey={5}>Aquisitions and Losses</NavItem>
-              </LinkContainer>
-              <LinkContainer to="/app/manage/logs">
-                <NavItem eventKey={6}>Logs</NavItem>
-              </LinkContainer>
-              <LinkContainer to="/app/manage/emails">
-                <NavItem eventKey={8}>Emails</NavItem>
-              </LinkContainer>
-            </Nav>
-          </Col>
-          <Col sm={10} >
+
+          <Col sm={12} >
             { this.props.children }
           </Col>
         </Row>
